Exercise 3.19 - 3.21 Add form validation and display errors on the UI
- Exercise 3.19 Validation
Expand the validation so that the name stored in the database has to
be at least three characters long.

Expand the frontend so that it displays some form of error message when
a validation error occurs.

Enable mongoose validators for Update operations

- Exercise 3.20
Add validation to ensure phone numbers are of the correct form.
A phone number must:

have length of 8 or more
be formed of two parts that are separated by -, the first part has two
or three numbers and the second part also consists of numbers

eg. 09-1234556 and 040-22334455 are valid phone numbers
eg. 1234556, 1-22334455 and 10-22-334455 are invalid

- Exercise 3.21
Redeploy web app